+++ zarrs/src/array/codec/array_to_bytes/vlen_v2/vlen_v2_codec.rs
@@ -93,6 +93,10 @@ impl UnboundArrayToBytesCodecTraits for VlenV2Codec {
 }
 
 impl ArrayCodecTraits for VlenV2CodecBound {
+    fn as_any(&self) -> &dyn std::any::Any {
+        self
+    }
+
     fn data_type(&self) -> &DataType {
         &self.data_type
     }